# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stores:
    get:
      operationId: listStores
      tags: [Stores]
      summary: List active (approved) stores
      responses:
        "200":
          description: Active stores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Store"
    post:
      operationId: createStore
      tags: [Stores]
      summary: Register a store (merchant) — created pending admin review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StoreInput"
      responses:
        "201":
          description: Created store (pending)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /stores/mine:
    get:
      operationId: listMyStores
      tags: [Stores]
      summary: List stores owned by the authenticated merchant
      responses:
        "200":
          description: The caller's stores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Store"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /stores/admin:
    get:
      operationId: listAllStores
      tags: [Stores]
      summary: List all stores in any status (admin only)
      responses:
        "200":
          description: All stores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Store"
        "401":
          description: Not authenticated as admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /stores/{id}:
    get:
      operationId: getStore
      tags: [Stores]
      summary: Get a single store
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: The store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    patch:
      operationId: updateStore
      tags: [Stores]
      summary: Update a store (owner or admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StoreUpdate"
      responses:
        "200":
          description: Updated store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"
        "401":
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    delete:
      operationId: deleteStore
      tags: [Stores]
      summary: Permanently delete a store and its products (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Not authenticated as admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /stores/{id}/review:
    patch:
      operationId: reviewStore
      tags: [Stores]
      summary: Approve, reject, suspend or reactivate a store (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StoreReviewInput"
      responses:
        "200":
          description: Reviewed store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"
        "401":
          description: Not authenticated as admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /stores/{id}/orders:
    get:
      operationId: listStoreOrders
      tags: [Stores]
      summary: List a store's orders (owner or admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: The store's orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
        "401":
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /stores/{id}/customers:
    get:
      operationId: listStoreCustomers
      tags: [Stores]
      summary: List customers who ordered from a store (owner or admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Aggregated customers for the store
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StoreCustomer"
        "401":
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /stores/{id}/refunds:
    get:
      operationId: listStoreRefunds
      tags: [Stores]
      summary: List compensation requests for a store's orders (owner or admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: The store's compensation requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StoreRefund"
        "401":
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /stores/{id}/refunds/{refundId}:
    patch:
      operationId: decideStoreRefund
      tags: [Stores]
      summary: Approve (with amount) or reject a compensation for the store's own order (owner or admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: refundId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefundDecisionInput"
      responses:
        "200":
          description: Updated refund
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Refund"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "409":
          description: Already reviewed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
components:
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Store:
      type: object
      required: [id, name, address, storeType, ownerPhone, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        address:
          type: string
        description:
          type: ["string", "null"]
        storeType:
          type: string
          description: Kind of shop, e.g. "خضار وفواكه", "بقالة", "لحوم".
        ownerPhone:
          type: string
        imageUrl:
          type: ["string", "null"]
        status:
          type: string
          description: "قيد المراجعة | مفعّل | مرفوض | موقوف مؤقتاً"
        subscriptionExpiresAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
    StoreInput:
      type: object
      required: [name, address, storeType]
      properties:
        name:
          type: string
          minLength: 1
        address:
          type: string
          minLength: 1
        storeType:
          type: string
          minLength: 1
        description:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
    StoreUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        address:
          type: string
          minLength: 1
        storeType:
          type: string
          minLength: 1
        description:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
    StoreReviewInput:
      type: object
      required: [action]
      properties:
        action:
          type: string
          enum: [approve, reject, suspend, reactivate]
        subscriptionMonths:
          type: integer
          minimum: 1
          maximum: 24
          description: Subscription length in months on approval (default 3).
        subscriptionMode:
          type: string
          enum: [extend, set]
          description: >-
            How to apply subscriptionMonths on approval. "extend" (default) adds
            the months onto the later of now or the current expiry (renewals
            stack). "set" replaces the expiry with exactly that many months from
            today — used to correct or reduce a subscription after a misclick.
...
    StoreCustomer:
      type: object
      required: [phone, orderCount, totalSpent]
      properties:
        phone:
          type: string
        name:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        orderCount:
          type: integer
        totalSpent:
          type: integer
        lastOrderAt:
          type: ["string", "null"]
          format: date-time
    StoreRefund:
      type: object
      required: [id, orderId, customerPhone, productName, imageUrl, amount, status]
      properties:
        id:
          type: integer
        orderId:
          type: integer
        customerPhone:
          type: string
        customerName:
          type: ["string", "null"]
        productName:
          type: string
        imageUrl:
          type: string
        amount:
          type: integer
        status:
          type: string
        createdAt:
          type: ["string", "null"]
          format: date-time
        reviewedAt:
          type: ["string", "null"]
          format: date-time
...
    OrderItemInput:
      type: object
      required: [id, name, price, unit, qty]
      properties:
        id:
          type: integer
        name:
          type: string
        price:
          type: integer
        unit:
          type: string
        qty:
          type: integer
          minimum: 1
...
    Order:
      type: object
      required:
        - id
        - customerPhone
        - items
        - subtotal
        - deliveryFee
        - deliveryType
        - discountApplied
        - pointsEarned
        - pointsRedeemed
        - total
        - status
        - createdAt
      properties:
        id:
          type: integer
        customerPhone:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemInput"
        subtotal:
          type: integer
        deliveryFee:
          type: integer
        deliveryType:
          type: string
          enum: [standard, express]
        discountApplied:
          type: integer
        pointsEarned:
          type: integer
        pointsRedeemed:
          type: integer
        redemptionType:
          type: ["string", "null"]
        note:
          type: ["string", "null"]
          description: Optional customer note attached to the order.
        total:
          type: integer
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        status:
          type: string
        pickupTime:
          type: ["string", "null"]
        walletApplied:
          type: integer
        storeId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
    Refund:
      type: object
      required: [id, orderId, customerPhone, productName, imageUrl, amount, status, createdAt]
      properties:
        id:
          type: integer
        orderId:
          type: integer
        customerPhone:
          type: string
        productName:
          type: string
        imageUrl:
          type: string
        amount:
          type: integer
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        reviewedAt:
          type: string
          format: date-time
          nullable: true
...
    RefundDecisionInput:
      type: object
      required: [action]
      properties:
        action:
          type: string
          enum: [approve, reject]
        amount:
          type: integer
          minimum: 0
...